refactor(admin): split monolith page into focused components with optimistic ux
- extract Stat, SectionTitle, AddressTable, OracleTable, RoleSelect,
  SolanaAdminPanel, QubicAdminPanel into separate files (admin.page.tsx
  872->149 lines)
- AddressTable/OracleTable: optimistically hide removed items on click,
  revert on tx failure
- QubicAdminPanel: optimistic oracle/pauser list updates after add/remove role
- all add/remove forms clear their inputs on success (via onSuccess callback
  added to AdminActionForm)
- oracle claim button styled consistently with other form submit buttons
- solana fee stats and claim protocol fee form laid out side-by-side (sm:grid-cols-2)
- added bottom padding to admin page
